gh-102371: move _Py_Mangle from compile.c to symtable.c (#102372)
